Fix serverless package breaking build
The serverless package must be installed globally. If it isn't then the
build will fail because the assembly is the same for the template
project as the actual application. Installing the deployment plugin in
the project is totally fine though.

I've updated the package.json and the README to reflect this